pyproject.toml
[project]
name = "gameserver"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi[all,standard]>=0.116.2",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.25",
]
